Move the code that performs the writing to lib/cli
The CLI is the only entity that calls it directly.
