+++ lib/cli/etcher.js
@@ -18,7 +18,7 @@
 
 const visuals = require('resin-cli-visuals');
 const form = require('resin-cli-form');
-const writer = require('../src/writer');
+const writer = require('./writer');
 const utils = require('./utils');
 const options = require('./cli');
 const EXIT_CODES = require('../src/exit-codes');
