+++ lib/cli/writer.js
